requirements.txt
pytesseract
opencv-python-headless 
pillow 
easyocr
openai==0.28
requests